Update README.md
What is claimed is:
A payment card for multiple accounts, the payment card comprising:
a card member;
a plurality of individual sockets coupled to the card member, each of the individual sockets configured to accept a single programmable chip;
 a plurality of programmable chips, at least one of which is releasably attached to the card member by way of one of the plurality of individual sockets; each programmable chip associated with a different payment account having a primary account number (PAN), and each programmable chip including a personal identification number (PIN) for the corresponding payment account associated with said programmable chip; and

a control chip coupled to each of the plurality of sockets and configured to select one of the programmable chips to be active during a financial transaction only after the PIN for the selected one of the programmable chips is provided by a user associated   with the payment card.

The payment card of claim 1, further comprising a magnetic stripe.  

The payment card of claim 2, wherein the control chip is configured to communicate with the magnetic stripe; and wherein the control chip is configured to control whether the magnetic stripe is utilized in the transaction.

The payment card of claim 1, wherein the control chip is configured to provide a virtual payment number in the transaction for a payment account associated with the selected one of the programmable chips, in place of a PAN for said payment account. The payment card of claim 1, further comprising a wireless communication module coupled to the control chip; and wherein the control chip is configured to communicate one of a PAN and a virtual payment number associated with the selected one of the programmable chips in 	connection with the transaction.  

The payment card of claim 1, further comprising a biometric reader coupled to the control chip, the biometric reader configured to read a biometric of the user and communicate the biometric to the control chip; and wherein the control chip is configured to match the received biometric to a stored biometric prior to per- mitting the transaction to be completed.

The payment card of claim 6, wherein the biometric reader is operable to read a fingerprint.  

The payment card of claim 1, wherein the selected one of the programmable chips includes an indicator for visually indicating the selected one of the programmable chips.  

The payment card of claim 1, further comprising a display. 

The payment card of claim 1, further comprising a power source.

A method of using a payment card for multiple accounts, the payment card comprising a card member, a plurality of individual sockets coupled to the card member, and a control chip coupled to each of the plurality of individual sockets, wherein each of the plurality of individual sockets is configured to releasably accept at least one programmable chip, the method comprising:

plugging at least one programmable chip into one of the plurality of individual sockets of the payment card, wherein the at least one programmable chip is associated with a payment account having a payment account number, and wherein the at least one programmable chip includes a personal identification number (PIN) for the corresponding payment account associated with said at least one programmable chip;

interfacing the payment card, as part of a financial trans- action, with a point of sale (POS) device and/or an automated teller machine (ATM); selecting, by the control chip, the at least one programmable chip for activation during the financial transaction;

receiving, by the control chip, the PIN for the corresponding payment account associated with said at least one programmable chip from the POS device and/or the ATM; and  

the ATM, the payment account number for the corresponding payment account associated with said at least one programmable chip only after the control chip verifies the received PIN.

10 12. The method of claim 11, wherein providing the payment account number to the POS device and/or the ATM includes providing a virtual payment number in place of a primary account number (PAN) for the corresponding payment account associated with said at least one program-

15 mable chip.

The method of claim 11, wherein the at least one programmable chip comprises an indicator; and

further comprising visually indicating, by the indicator, the at least one programmable chip when the at least programmable chip is selected for activation during the financial transaction.
The method of claim 11, further comprising deriving, by the payment card, power for the payment card from the POS device and/or the ATM when the payment card is
25 interfacing with the POS device and/or the ATM in connection with the transaction.
The method of claim 14, further comprising receiving a user input indicative of the at least one programmable chip, via the POS device and/or the ATM, prior to selecting the at least one programmable chip; and
wherein selecting the at least one programmable chip for activation includes selecting the at least one programmable chip for activation based on the user input.